% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/appengine_functions.R
\name{apps.get}
\alias{apps.get}
\title{Gets information about an application.}
\usage{
apps.get(appsId, ensureResourcesExist = NULL)
}
\arguments{
\item{appsId}{Part of `name`}

\item{ensureResourcesExist}{Certain resources associated with an application are created on-demand}
}
\description{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_skeleton}}
}
\details{
Authentication scopes used by this function are:
\itemize{
  \item https://www.googleapis.com/auth/cloud-platform
}

Set \code{options(googleAuthR.scopes.selected = c(https://www.googleapis.com/auth/cloud-platform)}
Then run \code{googleAuthR::gar_auth()} to authenticate.
See \code{\link[googleAuthR]{gar_auth}} for details.
}
\seealso{
\href{https://cloud.google.com/appengine/docs/admin-api/}{Google Documentation}
}
