% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/appengine_objects.R
\name{AutomaticScaling}
\alias{AutomaticScaling}
\title{AutomaticScaling Object}
\usage{
AutomaticScaling(coolDownPeriod = NULL, cpuUtilization = NULL,
  maxConcurrentRequests = NULL, maxIdleInstances = NULL,
  maxTotalInstances = NULL, maxPendingLatency = NULL,
  minIdleInstances = NULL, minTotalInstances = NULL,
  minPendingLatency = NULL, requestUtilization = NULL,
  diskUtilization = NULL, networkUtilization = NULL)
}
\arguments{
\item{coolDownPeriod}{Amount of time that the [Autoscaler](https://cloud}

\item{cpuUtilization}{Target scaling by CPU usage}

\item{maxConcurrentRequests}{Number of concurrent requests an automatic scaling instance can accept before the scheduler spawns a new instance}

\item{maxIdleInstances}{Maximum number of idle instances that should be maintained for this version}

\item{maxTotalInstances}{Maximum number of instances that should be started to handle requests}

\item{maxPendingLatency}{Maximum amount of time that a request should wait in the pending queue before starting a new instance to handle it}

\item{minIdleInstances}{Minimum number of idle instances that should be maintained for this version}

\item{minTotalInstances}{Minimum number of instances that should be maintained for this version}

\item{minPendingLatency}{Minimum amount of time a request should wait in the pending queue before starting a new instance to handle it}

\item{requestUtilization}{Target scaling by request utilization}

\item{diskUtilization}{Target scaling by disk usage}

\item{networkUtilization}{Target scaling by network usage}
}
\value{
AutomaticScaling object
}
\description{
AutomaticScaling Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
Automatic scaling is based on request rate, response latencies, and other application metrics.
}

