% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/appengine_objects.R
\name{Library}
\alias{Library}
\title{Library Object}
\usage{
Library(name = NULL, version = NULL)
}
\arguments{
\item{name}{Name of the library}

\item{version}{Version of the library to select, or 'latest'}
}
\value{
Library object
}
\description{
Library Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
Third-party Python runtime library that is required by the application.
}

