% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/appengine_objects.R
\name{ListModulesResponse}
\alias{ListModulesResponse}
\title{ListModulesResponse Object}
\usage{
ListModulesResponse(modules = NULL, nextPageToken = NULL)
}
\arguments{
\item{modules}{The modules belonging to the requested application}

\item{nextPageToken}{Continuation token for fetching the next page of results}
}
\value{
ListModulesResponse object
}
\description{
ListModulesResponse Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
Response message for Modules.ListModules.
}
