% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/appengine_objects.R
\name{ManualScaling}
\alias{ManualScaling}
\title{ManualScaling Object}
\usage{
ManualScaling(instances = NULL)
}
\arguments{
\item{instances}{Number of instances to assign to the module at the start}
}
\value{
ManualScaling object
}
\description{
ManualScaling Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
A module with manual scaling runs continuously, allowing you to perform complex initialization and rely on the state of its memory over time.
}

