% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/appengine_objects.R
\name{Module}
\alias{Module}
\title{Module Object}
\usage{
Module(name = NULL, id = NULL, split = NULL)
}
\arguments{
\item{name}{Full path to the Module resource in the API}

\item{id}{Relative name of the module within the application}

\item{split}{Mapping that defines fractional HTTP traffic diversion to different versions within the module}
}
\value{
Module object
}
\description{
Module Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
A Module resource is a logical component of an application that can share state and communicate in a secure fashion with other modules. For example, an application that handles customer requests might include separate modules to handle tasks such as backend data analysis or API requests from mobile devices. Each module has a collection of versions that define a specific set of code used to implement the functionality of that module.
}
