% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/appengine_objects.R
\name{Network}
\alias{Network}
\title{Network Object}
\usage{
Network(forwardedPorts = NULL, instanceTag = NULL, name = NULL)
}
\arguments{
\item{forwardedPorts}{List of ports, or port pairs, to forward from the virtual machine to the application container}

\item{instanceTag}{Tag to apply to the VM instance during creation}

\item{name}{Google Cloud Platform network where the virtual machines are created}
}
\value{
Network object
}
\description{
Network Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
Extra network settings. Only applicable for VM runtimes.
}
