% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/appengine_objects.R
\name{ScriptHandler}
\alias{ScriptHandler}
\title{ScriptHandler Object}
\usage{
ScriptHandler(scriptPath = NULL)
}
\arguments{
\item{scriptPath}{Path to the script from the application root directory}
}
\value{
ScriptHandler object
}
\description{
ScriptHandler Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
Executes a script to handle the request that matches the URL pattern.
}

