% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/appengine_objects.R
\name{StaticDirectoryHandler.httpHeaders}
\alias{StaticDirectoryHandler.httpHeaders}
\title{StaticDirectoryHandler.httpHeaders Object}
\usage{
StaticDirectoryHandler.httpHeaders()
}
\value{
StaticDirectoryHandler.httpHeaders object
}
\description{
StaticDirectoryHandler.httpHeaders Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
HTTP headers to use for all responses from these URLs.
}
\seealso{
Other StaticDirectoryHandler functions: \code{\link{StaticDirectoryHandler}}
}

