% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/appengine_objects.R
\name{StaticDirectoryHandler}
\alias{StaticDirectoryHandler}
\title{StaticDirectoryHandler Object}
\usage{
StaticDirectoryHandler(StaticDirectoryHandler.httpHeaders = NULL,
  directory = NULL, httpHeaders = NULL, mimeType = NULL,
  expiration = NULL, requireMatchingFile = NULL,
  applicationReadable = NULL)
}
\arguments{
\item{StaticDirectoryHandler.httpHeaders}{The \link{StaticDirectoryHandler.httpHeaders} object or list of objects}

\item{directory}{Path to the directory containing the static files from the application root directory}

\item{httpHeaders}{HTTP headers to use for all responses from these URLs}

\item{mimeType}{MIME type used to serve all files served by this handler}

\item{expiration}{Time a static file served by this handler should be cached}

\item{requireMatchingFile}{Whether this handler should match the request if the file referenced by the handler does not exist}

\item{applicationReadable}{Whether files should also be uploaded as code data}
}
\value{
StaticDirectoryHandler object
}
\description{
StaticDirectoryHandler Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
Files served directly to the user for a given URL, such as images, CSS stylesheets, or JavaScript source files. Static directory handlers make it easy to serve the entire contents of a directory as static files.
}
\seealso{
Other StaticDirectoryHandler functions: \code{\link{StaticDirectoryHandler.httpHeaders}}
}

