% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/appengine_objects.R
\name{StaticFilesHandler.httpHeaders}
\alias{StaticFilesHandler.httpHeaders}
\title{StaticFilesHandler.httpHeaders Object}
\usage{
StaticFilesHandler.httpHeaders()
}
\value{
StaticFilesHandler.httpHeaders object
}
\description{
StaticFilesHandler.httpHeaders Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
HTTP headers to use for all responses from these URLs.
}
\seealso{
Other StaticFilesHandler functions: \code{\link{StaticFilesHandler}}
}

