% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/appengine_objects.R
\name{TrafficSplit}
\alias{TrafficSplit}
\title{TrafficSplit Object}
\usage{
TrafficSplit(TrafficSplit.allocations = NULL, shardBy = NULL,
  allocations = NULL)
}
\arguments{
\item{TrafficSplit.allocations}{The \link{TrafficSplit.allocations} object or list of objects}

\item{shardBy}{Mechanism used to determine which version a request is sent to}

\item{allocations}{Mapping from version IDs within the module to fractional (0}
}
\value{
TrafficSplit object
}
\description{
TrafficSplit Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
Traffic routing configuration for versions within a single module. Traffic splits define how traffic directed to the module is assigned to versions.
}
\seealso{
Other TrafficSplit functions: \code{\link{TrafficSplit.allocations}}
}
