% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/appengine_objects.R
\name{Version.envVariables}
\alias{Version.envVariables}
\title{Version.envVariables Object}
\usage{
Version.envVariables()
}
\value{
Version.envVariables object
}
\description{
Version.envVariables Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
Environment variables made available to the application. Only returned in `GET` requests if `view=FULL` is set.
}
\seealso{
Other Version functions: \code{\link{Version.betaSettings}},
  \code{\link{Version}}
}

