% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/appengine_objects.R
\name{Volume}
\alias{Volume}
\title{Volume Object}
\usage{
Volume(name = NULL, volumeType = NULL, sizeGb = NULL)
}
\arguments{
\item{name}{Unique name for the volume}

\item{volumeType}{Underlying volume type, e}

\item{sizeGb}{Volume size in gigabytes}
}
\value{
Volume object
}
\description{
Volume Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
Volumes mounted within the app container. Only applicable for VM runtimes.
}
