% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/appengine_objects.R
\name{Application}
\alias{Application}
\title{Google App Engine Admin API Objects 
Provisions and manages App Engine applications.}
\usage{
Application(name = NULL, id = NULL, dispatchRules = NULL,
  authDomain = NULL, location = NULL, codeBucket = NULL,
  defaultCookieExpiration = NULL, defaultHostname = NULL,
  defaultBucket = NULL)
}
\arguments{
\item{name}{Full path to the Application resource in the API}

\item{id}{Identifier of the Application resource}

\item{dispatchRules}{HTTP path dispatch rules for requests to the application that do not explicitly target a module or version}

\item{authDomain}{Google Apps authentication domain that controls which users can access this application}

\item{location}{Location from which this application will be run}

\item{codeBucket}{Google Cloud Storage bucket that can be used for storing files associated with this application}

\item{defaultCookieExpiration}{Cookie expiration policy for this application}

\item{defaultHostname}{Hostname used to reach the application, as resolved by App Engine}

\item{defaultBucket}{Google Cloud Storage bucket that can be used by this application to store content}
}
\value{
Application object
}
\description{
Auto-generated code by googleAuthR::gar_create_api_objects
 at 2016-09-03 23:15:57
filename: /Users/mark/dev/R/autoGoogleAPI/googleappenginev1beta4.auto/R/appengine_objects.R
api_json: api_json
}
\details{
Objects for use by the functions created by googleAuthR::gar_create_api_skeleton
Application Object


Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
An Application resource contains the top-level configuration of an App Engine application.
}
\seealso{
Other Application functions: \code{\link{apps.create}}
}

